ICSE2022ReplicationPackage/requirements.txt
ansicolors==1.1.8
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
et-xmlfile==1.1.0
Flask==2.0.1
Flask-WTF==0.15.1
gitdb==4.0.7
GitPython==3.1.18
idna==3.2
ijson==3.1.4
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonpickle==2.0.0
lizard==1.17.9
MarkupSafe==2.0.1
openpyxl==3.0.7
PyDriller==1.15.5
pymongo==3.12.0
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
smmap==4.0.0
tqdm==4.62.2
typing-extensions==3.10.0.0
unidiff==0.6.0
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.5.0
